BehaviorTest: speed input tests
Configure permission once and then don't wait for it.
