+++ app/src/androidTest/java/page/ooooo/geoshare/BehaviorTest.kt
@@ -29,6 +29,7 @@ import org.junit.AssumptionViolatedException
 import page.ooooo.geoshare.data.local.database.Server
 import page.ooooo.geoshare.data.local.database.ServerAuthType
 import page.ooooo.geoshare.data.local.preferences.CoordinateFormat
+import page.ooooo.geoshare.data.local.preferences.Permission
 import page.ooooo.geoshare.lib.android.PackageNames
 import page.ooooo.geoshare.lib.calcExponentialBackoffMillis
 import page.ooooo.geoshare.lib.formatters.CoordinateFormatter
@@ -377,6 +378,96 @@ fun UiAutomatorTestScope.shareUri(unsafeUriString: String) {
         @Suppress("SpellCheckingInspection") "am start -a android.intent.action.VIEW -d $unsafeUriString -n ${BuildConfig.APPLICATION_ID}/page.ooooo.geoshare.ConversionActivity ${BuildConfig.APPLICATION_ID}"
     )
 }
+
+fun UiAutomatorTestScope.configureConnectionPermissionPreference(permission: Permission) {
+    goToUserPreferencesDetail(UserPreferenceGroupId.CONNECTION_PERMISSION)
+    onElement { viewIdResourceName == "geoShareUserPreferenceConnectionPermission_$permission" }.click()
+}
+
+fun UiAutomatorTestScope.testUri(
+    expectedPoints: Points,
+    unsafeUriString: String,
+    fallbackPoints: Points? = null,
+    accurate: Boolean? = null,
+    timeoutMs: Long = NETWORK_TIMEOUT,
+) {
+    shareUri(unsafeUriString)
+    quickWaitForStableInActiveWindow() // Wait for the result to render, because there might be the old result
+    try {
+        assertConversionSucceeds(expectedPoints, accurate, timeoutMs)
+    } catch (e: AssertionError) {
+        if (fallbackPoints != null) {
+            assertConversionSucceeds(fallbackPoints, accurate, timeoutMs)
+        } else {
+            throw e
+        }
+    }
+}
+
+fun UiAutomatorTestScope.testUri(
+    expectedPoint: Point,
+    unsafeUriString: String,
+    fallbackPoint: Point? = null,
+    accurate: Boolean? = null,
+    timeoutMs: Long = NETWORK_TIMEOUT,
+) =
+    testUri(
+        persistentListOf(expectedPoint),
+        unsafeUriString,
+        fallbackPoint?.let { persistentListOf(it) },
+        accurate,
+        timeoutMs,
+    )
+
+fun UiAutomatorTestScope.testUriFails(
+    expectedMessage: Set<String>,
+    unsafeUriString: String,
+    timeoutMs: Long = NETWORK_TIMEOUT,
+) {
+    shareUri(unsafeUriString)
+    quickWaitForStableInActiveWindow() // Wait for the result to render, because there might be the old result
+    assertConversionFails(expectedMessage, timeoutMs)
+}
+
+fun UiAutomatorTestScope.testUriAnyCoordinates(
+    unsafeUriString: String,
+    timeoutMs: Long = NETWORK_TIMEOUT,
+) {
+    shareUri(unsafeUriString)
+    quickWaitForStableInActiveWindow() // Wait for the result to render, because there might be the old result
+    assertConversionSucceedsAnyCoordinates(timeoutMs)
+}
+
+fun UiAutomatorTestScope.testText(expectedPoints: Points, unsafeText: String) {
+    // It would be preferable to test sharing of the text with the app, but this shell command doesn't work when
+    // there are spaces in the text. So instead, we type the text in the main form of the app.
+    // device.executeShellCommand(
+    //     "am start -a android.intent.action.SEND -t text/plain -e android.intent.extra.TEXT $unsafeText -n ${BuildConfig.APPLICATION_ID}.debug/${BuildConfig.APPLICATION_ID}.ConversionActivity ${BuildConfig.APPLICATION_ID}.debug"
+    // )
+
+    // Go to main form, if we're on the result screen
+    onElementOrNull(1_000L) { viewIdResourceName == "geoShareMainBackButton" }?.click()
+
+    // Set main input
+    val mainInput = onElement { viewIdResourceName == "geoShareMainSourceTextField" }
+    mainInput.setText(unsafeText)
+    quickWaitForStableInActiveWindow() // Wait for the submit button to get its final position, after setting text
+
+    // Submit main form
+    if (mainInput.isFocused) {
+        // If the field is focused, the submit button can be covered by IME, so submit by pressing Enter
+        pressEnter()
+    } else {
+        // If the field is not focused, then pressing Enter doesn't submit, so submit by clicking the submit button
+        onElement { viewIdResourceName == "geoShareMainSubmitButton" }.click()
+    }
+
+    // Shows points
+    assertConversionSucceeds(expectedPoints)
+}
+
+fun UiAutomatorTestScope.testText(expectedPoint: Point, unsafeText: String) =
+    testText(persistentListOf(expectedPoint), unsafeText)
 
 fun UiAutomatorTestScope.goToInputList() {
     // If we're on the main screen, use the main menu
