+++ app/src/androidTest/java/page/ooooo/geoshare/MainBehaviorTest.kt
@@ -2,7 +2,6 @@ package page.ooooo.geoshare
 
 import androidx.test.uiautomator.uiAutomator
 import org.junit.Test
-import page.ooooo.geoshare.inputs.testText
 import page.ooooo.geoshare.lib.android.PackageNames
 import page.ooooo.geoshare.lib.geo.Source
 import page.ooooo.geoshare.lib.geo.WGS84Point
