+++ app/src/androidTest/java/page/ooooo/geoshare/inputs/AppleMapsInputBehaviorTest.kt
@@ -5,10 +5,13 @@ import kotlinx.coroutines.runBlocking
 import org.junit.Test
 import page.ooooo.geoshare.assumeDomainResolvable
 import page.ooooo.geoshare.closeIntro
+import page.ooooo.geoshare.configureConnectionPermissionPreference
 import page.ooooo.geoshare.data.local.preferences.Permission
 import page.ooooo.geoshare.launchApplication
 import page.ooooo.geoshare.lib.geo.Source
 import page.ooooo.geoshare.lib.geo.WGS84Point
+import page.ooooo.geoshare.testText
+import page.ooooo.geoshare.testUri
 import page.ooooo.geoshare.waitForAppToBeVisible
 
 class AppleMapsInputBehaviorTest {
