+++ app/src/androidTest/java/page/ooooo/geoshare/inputs/BaiduMapInputBehaviorTest.kt
@@ -6,9 +6,15 @@ import kotlinx.coroutines.runBlocking
 import org.junit.Test
 import page.ooooo.geoshare.NETWORK_TIMEOUT
 import page.ooooo.geoshare.assumeDomainResolvable
+import page.ooooo.geoshare.closeIntro
+import page.ooooo.geoshare.configureConnectionPermissionPreference
+import page.ooooo.geoshare.data.local.preferences.Permission
+import page.ooooo.geoshare.launchApplication
 import page.ooooo.geoshare.lib.geo.BD09MCPoint
 import page.ooooo.geoshare.lib.geo.GCJ02Point
 import page.ooooo.geoshare.lib.geo.Source
+import page.ooooo.geoshare.testUri
+import page.ooooo.geoshare.waitForAppToBeVisible
 
 class BaiduMapInputBehaviorTest {
     @Test
@@ -100,6 +106,12 @@ class BaiduMapInputBehaviorTest {
         runBlocking {
             assumeDomainResolvable("map.baidu.com")
         }
+
+        // Launch app and close intro
+        launchApplication()
+        waitForAppToBeVisible()
+        closeIntro()
+        configureConnectionPermissionPreference(Permission.ALWAYS)
 
         // Shared coordinates
         testUri(
