--- app/src/androidTest/java/page/ooooo/geoshare/inputs/InputBehaviorTest.kt
@@ -1,132 +1,0 @@
-package page.ooooo.geoshare.inputs
-
-import androidx.test.uiautomator.UiAutomatorTestScope
-import kotlinx.collections.immutable.persistentListOf
-import page.ooooo.geoshare.NETWORK_TIMEOUT
-import page.ooooo.geoshare.assertConversionFails
-import page.ooooo.geoshare.assertConversionSucceeds
-import page.ooooo.geoshare.assertConversionSucceedsAnyCoordinates
-import page.ooooo.geoshare.data.local.preferences.Permission
-import page.ooooo.geoshare.goToUserPreferencesDetail
-import page.ooooo.geoshare.lib.geo.Point
-import page.ooooo.geoshare.lib.geo.Points
-import page.ooooo.geoshare.quickWaitForStableInActiveWindow
-import page.ooooo.geoshare.shareUri
-import page.ooooo.geoshare.ui.UserPreferenceGroupId
-
-fun UiAutomatorTestScope.configureConnectionPermissionPreference(permission: Permission) {
-    goToUserPreferencesDetail(UserPreferenceGroupId.CONNECTION_PERMISSION)
-    onElement { viewIdResourceName == "geoShareUserPreferenceConnectionPermission_$permission" }.click()
-}
-
-fun UiAutomatorTestScope.confirmDialogIfVisible() {
-    onElementOrNull(3_000L) { viewIdResourceName == "geoShareConfirmationDialogConfirmButton" }?.click()
-}
-
-fun UiAutomatorTestScope.goToMainForm() {
-    // Make sure we leave the result screen, if we're there, so that we don't accidentally test the old result.
-    onElementOrNull(1_000L) { viewIdResourceName == "geoShareMainBackButton" }?.click()
-}
-
-fun UiAutomatorTestScope.testUri(
-    expectedPoints: Points,
-    unsafeUriString: String,
-    fallbackPoints: Points? = null,
-    accurate: Boolean? = null,
-    timeoutMs: Long = NETWORK_TIMEOUT,
-) {
-    // Go to main form
-    goToMainForm()
-
-    // Share URI and confirm permission dialog
-    shareUri(unsafeUriString)
-    confirmDialogIfVisible()
-
-    try {
-        assertConversionSucceeds(expectedPoints, accurate, timeoutMs)
-    } catch (e: AssertionError) {
-        if (fallbackPoints != null) {
-            assertConversionSucceeds(fallbackPoints, accurate, timeoutMs)
-        } else {
-            throw e
-        }
-    }
-}
-
-fun UiAutomatorTestScope.testUri(
-    expectedPoint: Point,
-    unsafeUriString: String,
-    fallbackPoint: Point? = null,
-    accurate: Boolean? = null,
-    timeoutMs: Long = NETWORK_TIMEOUT,
-) =
-    testUri(
-        persistentListOf(expectedPoint),
-        unsafeUriString,
-        fallbackPoint?.let { persistentListOf(it) },
-        accurate,
-        timeoutMs,
-    )
-
-fun UiAutomatorTestScope.testUriFails(
-    expectedMessage: Set<String>,
-    unsafeUriString: String,
-    timeoutMs: Long = NETWORK_TIMEOUT,
-) {
-    // Go to main form
-    goToMainForm()
-
-    // Share URI and confirm permission dialog
-    shareUri(unsafeUriString)
-    confirmDialogIfVisible()
-
-    assertConversionFails(expectedMessage, timeoutMs)
-}
-
-fun UiAutomatorTestScope.testUriAnyCoordinates(
-    unsafeUriString: String,
-    timeoutMs: Long = NETWORK_TIMEOUT,
-) {
-    // Go to main form
-    goToMainForm()
-
-    // Share URI and confirm permission dialog
-    shareUri(unsafeUriString)
-    confirmDialogIfVisible()
-
-    assertConversionSucceedsAnyCoordinates(timeoutMs)
-}
-
-fun UiAutomatorTestScope.testText(expectedPoints: Points, unsafeText: String) {
-    // It would be preferable to test sharing of the text with the app, but this shell command doesn't work when
-    // there are spaces in the text. So instead, we type the text in the main form of the app.
-    // device.executeShellCommand(
-    //     "am start -a android.intent.action.SEND -t text/plain -e android.intent.extra.TEXT $unsafeText -n ${BuildConfig.APPLICATION_ID}.debug/${BuildConfig.APPLICATION_ID}.ConversionActivity ${BuildConfig.APPLICATION_ID}.debug"
-    // )
-
-    // Go to main form
-    goToMainForm()
-
-    // Set main input
-    val mainInput = onElement { viewIdResourceName == "geoShareMainSourceTextField" }
-    mainInput.setText(unsafeText)
-    quickWaitForStableInActiveWindow() // Wait for the submit button to get its final position, after setting text
-
-    // Submit main form
-    if (mainInput.isFocused) {
-        // If the field is focused, the submit button can be covered by IME, so submit by pressing Enter
-        pressEnter()
-    } else {
-        // If the field is not focused, then pressing Enter doesn't submit, so submit by clicking the submit button
-        onElement { viewIdResourceName == "geoShareMainSubmitButton" }.click()
-    }
-
-    // Confirm permission dialog
-    confirmDialogIfVisible()
-
-    // Shows points
-    assertConversionSucceeds(expectedPoints)
-}
-
-fun UiAutomatorTestScope.testText(expectedPoint: Point, unsafeText: String) =
-    testText(persistentListOf(expectedPoint), unsafeText)
