+++ app/src/androidTest/java/page/ooooo/geoshare/inputs/MapyComInputBehaviorTest.kt
@@ -4,8 +4,14 @@ import androidx.test.uiautomator.uiAutomator
 import kotlinx.coroutines.runBlocking
 import org.junit.Test
 import page.ooooo.geoshare.assumeDomainResolvable
+import page.ooooo.geoshare.closeIntro
+import page.ooooo.geoshare.configureConnectionPermissionPreference
+import page.ooooo.geoshare.data.local.preferences.Permission
+import page.ooooo.geoshare.launchApplication
 import page.ooooo.geoshare.lib.geo.Source
 import page.ooooo.geoshare.lib.geo.WGS84Point
+import page.ooooo.geoshare.testUri
+import page.ooooo.geoshare.waitForAppToBeVisible
 
 class MapyComInputBehaviorTest {
     @Test
@@ -32,6 +38,12 @@ class MapyComInputBehaviorTest {
         runBlocking {
             assumeDomainResolvable("mapy.com")
         }
+
+        // Launch app and close intro
+        launchApplication()
+        waitForAppToBeVisible()
+        closeIntro()
+        configureConnectionPermissionPreference(Permission.ALWAYS)
 
         // Short link
         testUri(
