+++ app/src/androidTest/java/page/ooooo/geoshare/inputs/UrbiInputBehaviorTest.kt
@@ -4,8 +4,14 @@ import androidx.test.uiautomator.uiAutomator
 import kotlinx.coroutines.runBlocking
 import org.junit.Test
 import page.ooooo.geoshare.assumeDomainResolvable
+import page.ooooo.geoshare.closeIntro
+import page.ooooo.geoshare.configureConnectionPermissionPreference
+import page.ooooo.geoshare.data.local.preferences.Permission
+import page.ooooo.geoshare.launchApplication
 import page.ooooo.geoshare.lib.geo.Source
 import page.ooooo.geoshare.lib.geo.WGS84Point
+import page.ooooo.geoshare.testUri
+import page.ooooo.geoshare.waitForAppToBeVisible
 
 class UrbiInputBehaviorTest {
     @Test
@@ -22,6 +28,12 @@ class UrbiInputBehaviorTest {
         runBlocking {
             assumeDomainResolvable("go.2gis.com")
         }
+
+        // Launch app and close intro
+        launchApplication()
+        waitForAppToBeVisible()
+        closeIntro()
+        configureConnectionPermissionPreference(Permission.ALWAYS)
 
         // Short link
         testUri(
