+++ app/src/androidTest/java/page/ooooo/geoshare/inputs/WazeInputBehaviorTest.kt
@@ -5,9 +5,13 @@ import kotlinx.coroutines.runBlocking
 import org.junit.Test
 import page.ooooo.geoshare.assumeDomainResolvable
 import page.ooooo.geoshare.closeIntro
+import page.ooooo.geoshare.configureConnectionPermissionPreference
+import page.ooooo.geoshare.data.local.preferences.Permission
 import page.ooooo.geoshare.launchApplication
 import page.ooooo.geoshare.lib.geo.Source
 import page.ooooo.geoshare.lib.geo.WGS84Point
+import page.ooooo.geoshare.testText
+import page.ooooo.geoshare.testUri
 import page.ooooo.geoshare.waitForAppToBeVisible
 
 class WazeInputBehaviorTest {
@@ -53,6 +57,12 @@ class WazeInputBehaviorTest {
         runBlocking {
             assumeDomainResolvable("waze.com")
         }
+
+        // Launch app and close intro
+        launchApplication()
+        waitForAppToBeVisible()
+        closeIntro()
+        configureConnectionPermissionPreference(Permission.ALWAYS)
 
         // Place id
         testUri(
